Move vector_db inside ML
