+++ mobile/apps/photos/rust/src/api/usearch_api.rs
@@ -1,4 +1,4 @@
-use ente_photos::vector_db;
+use ente_photos::ml::vector_db;
 use flutter_rust_bridge::frb;
 
 type SearchMatch = (Vec<u64>, Vec<f32>);
