+++ rust/photos/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod ml;
-pub mod vector_db;
 
 use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
