+++ rust/photos/src/ml/mod.rs
@@ -7,3 +7,4 @@ pub mod pet;
 pub mod preprocess;
 pub mod runtime;
 pub mod types;
+pub mod vector_db;
