+++ rust/photos/src/ml/vector_db.rs
